requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
flash-attn>=2.0.0; platform_system != "Darwin" or platform_machine != "arm64"
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.2.0

# Data processing
datasets>=2.12.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.12.0

# Configuration
pyyaml>=6.0.0

# Server and API
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.30.0
pydantic>=1.10.0

# UI
streamlit>=1.22.0

# Utils
python-dotenv>=1.0.0
